feat: add a bundler system to bundle php files
Since php does not come with a bundler and our goal is to make
development easier. I want to make a bundler where I can direct which
files are needed and break up the onefilecms.php into modular functions.

This will be used to try and differentiate the functions used in the php
file only to help me understand.